rasterizer_cache: Move utlities and type to separate headers
