+++ src/video_core/CMakeLists.txt
@@ -27,6 +27,9 @@ add_library(video_core STATIC
     rasterizer_cache/pixel_format.h
     rasterizer_cache/rasterizer_cache.cpp
     rasterizer_cache/rasterizer_cache.h
+    rasterizer_cache/rasterizer_cache_types.h
+    rasterizer_cache/rasterizer_cache_utils.cpp
+    rasterizer_cache/rasterizer_cache_utils.h
     rasterizer_cache/surface_params.cpp
     rasterizer_cache/surface_params.h
     renderer_opengl/frame_dumper_opengl.cpp
