+++ src/video_core/rasterizer_cache/rasterizer_cache.cpp
@@ -38,50 +38,6 @@
 
 namespace OpenGL {
 
-constexpr FormatTuple tex_tuple = {GL_RGBA8, GL_RGBA, GL_UNSIGNED_BYTE};
-
-static constexpr std::array<FormatTuple, 4> depth_format_tuples = {{
-    {GL_DEPTH_COMPONENT16, GL_DEPTH_COMPONENT, GL_UNSIGNED_SHORT}, // D16
-    {},
-    {GL_DEPTH_COMPONENT24, GL_DEPTH_COMPONENT, GL_UNSIGNED_INT},   // D24
-    {GL_DEPTH24_STENCIL8, GL_DEPTH_STENCIL, GL_UNSIGNED_INT_24_8}, // D24S8
-}};
-
-static constexpr std::array<FormatTuple, 5> fb_format_tuples = {{
-    {GL_RGBA8, GL_RGBA, GL_UNSIGNED_INT_8_8_8_8},     // RGBA8
-    {GL_RGB8, GL_BGR, GL_UNSIGNED_BYTE},              // RGB8
-    {GL_RGB5_A1, GL_RGBA, GL_UNSIGNED_SHORT_5_5_5_1}, // RGB5A1
-    {GL_RGB565, GL_RGB, GL_UNSIGNED_SHORT_5_6_5},     // RGB565
-    {GL_RGBA4, GL_RGBA, GL_UNSIGNED_SHORT_4_4_4_4},   // RGBA4
-}};
-
-// Same as above, with minor changes for OpenGL ES. Replaced
-// GL_UNSIGNED_INT_8_8_8_8 with GL_UNSIGNED_BYTE and
-// GL_BGR with GL_RGB
-static constexpr std::array<FormatTuple, 5> fb_format_tuples_oes = {{
-    {GL_RGBA8, GL_RGBA, GL_UNSIGNED_BYTE},            // RGBA8
-    {GL_RGB8, GL_RGB, GL_UNSIGNED_BYTE},              // RGB8
-    {GL_RGB5_A1, GL_RGBA, GL_UNSIGNED_SHORT_5_5_5_1}, // RGB5A1
-    {GL_RGB565, GL_RGB, GL_UNSIGNED_SHORT_5_6_5},     // RGB565
-    {GL_RGBA4, GL_RGBA, GL_UNSIGNED_SHORT_4_4_4_4},   // RGBA4
-}};
-
-const FormatTuple& GetFormatTuple(PixelFormat pixel_format) {
-    const SurfaceType type = GetFormatType(pixel_format);
-    const std::size_t format_index = static_cast<std::size_t>(pixel_format);
-
-    if (type == SurfaceType::Color) {
-        ASSERT(format_index < fb_format_tuples.size());
-        return (GLES ? fb_format_tuples_oes : fb_format_tuples)[format_index];
-    } else if (type == SurfaceType::Depth || type == SurfaceType::DepthStencil) {
-        const std::size_t tuple_idx = format_index - 14;
-        ASSERT(tuple_idx < depth_format_tuples.size());
-        return depth_format_tuples[tuple_idx];
-    }
-
-    return tex_tuple;
-}
-
 template <typename Map, typename Interval>
 static constexpr auto RangeFromInterval(Map& map, const Interval& interval) {
     return boost::make_iterator_range(map.equal_range(interval));
